Add signal mask

+++ inc/slog_port.h
@@ -15,7 +15,7 @@ int slog_port_init(void);
 
 void slog_port_deinit(void);
 
-void slog_port_output(int level, const char *log, size_t size);
+void slog_port_output(uint8_t level, const char *log, size_t size);
 
 
 #endif  /* __SLOG_PORT_H */
